split test into own classes
