+++ UniTAS/Plugin.Tests/MovieRunner/CoroutineTests.cs
@@ -1,0 +1,57 @@
+namespace UniTAS.Plugin.Tests.MovieRunner;
+
+public class CoroutineTests
+{
+    [Fact]
+    public void MainCoroutine()
+    {
+        const string input = @"
+i = 0
+coroutine.yield()
+i = i + 1
+coroutine.yield()
+i = i + 1
+";
+        var movieRunner = Utils.Setup(input).Item1;
+
+        Assert.False(movieRunner.Finished);
+        movieRunner.Update();
+        Assert.False(movieRunner.Finished);
+        movieRunner.Update();
+        Assert.False(movieRunner.Finished);
+        movieRunner.Update();
+        Assert.True(movieRunner.Finished);
+    }
+
+    [Fact]
+    public void CoroutineWithoutYield()
+    {
+        const string input = @"
+i = 0
+i = i + 1
+";
+        var movieRunner = Utils.Setup(input).Item1;
+
+        Assert.False(movieRunner.Finished);
+        movieRunner.Update();
+        Assert.True(movieRunner.Finished);
+    }
+
+    [Fact]
+    public void CoroutineAlias()
+    {
+        const string input = @"
+i = 0
+adv()
+i = i + 1
+";
+
+        var movieRunner = Utils.Setup(input).Item1;
+
+        Assert.False(movieRunner.Finished);
+        movieRunner.Update();
+        Assert.False(movieRunner.Finished);
+        movieRunner.Update();
+        Assert.True(movieRunner.Finished);
+    }
+}
